fix : cache total dataset
