+++ client/utils/__init__.py
@@ -1,1 +1,2 @@
 from .api import APIHelper
+from .data import load_demo_data, load_total_data, download_data
